adding shared ribbon bar
